+++ CHANGELOG.md
@@ -5,6 +5,7 @@ xx.xx.xxxx
 
 - Export to Eclipse Vorto, first version
 - First version of export to SVG in oneM2M resource format
+- First version of export to oneM2M XSD 
 - Fixed errors in Java export
 - Fixed errors in OPML export
 - Improved markdown export